requirements.txt
# AlignAI Backend Requirements
# ==============================

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security
Werkzeug==3.0.1

# File Parsing
pypdf==3.17.1
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# AI/LLM (Optional - install if using AI features)
# openai==1.3.0
# anthropic==0.7.0
# langchain==0.1.0
